api_pipe/requirements.txt
Django==2.2.6
djangorestframework==3.10.3
imutils==0.5.3
numpy==1.17.3
opencv-python==4.1.1.26
pkg-resources==0.0.0
Pygments==2.4.2
pytz==2019.3
sqlparse==0.3.0
